Refactor organization creation dialog component
