+++ web/src/components/ui/dialog.tsx
@@ -63,7 +63,7 @@ function DialogContent({
                         render={
                             <Button
                                 variant="ghost"
-                                className="absolute top-2 right-2"
+                                className="absolute top-2 right-2 cursor-pointer"
                                 size="icon-sm"
                             />
                         }
